+++ commercetools/commercetools-sdk-java-api/src/main/java-generated/com/commercetools/api/models/recurring_order/RecurringOrderDraftBuilder.java
@@ -36,6 +36,9 @@ public class RecurringOrderDraftBuilder implements Builder<RecurringOrderDraft> 
     private java.time.ZonedDateTime startsAt;
 
     @Nullable
+    private java.time.ZonedDateTime expiresAt;
+
+    @Nullable
     private com.commercetools.api.models.state.StateResourceIdentifier state;
 
     @Nullable
@@ -106,6 +109,17 @@ public class RecurringOrderDraftBuilder implements Builder<RecurringOrderDraft> 
 
     public RecurringOrderDraftBuilder startsAt(final java.time.ZonedDateTime startsAt) {
         this.startsAt = startsAt;
+        return this;
+    }
+
+    /**
+     *  <p>Date and time (UTC) when the RecurringOrder will expire.</p>
+     * @param expiresAt value to be set
+     * @return Builder
+     */
+
+    public RecurringOrderDraftBuilder expiresAt(@Nullable final java.time.ZonedDateTime expiresAt) {
+        this.expiresAt = expiresAt;
         return this;
     }
 
@@ -219,6 +233,16 @@ public class RecurringOrderDraftBuilder implements Builder<RecurringOrderDraft> 
     }
 
     /**
+     *  <p>Date and time (UTC) when the RecurringOrder will expire.</p>
+     * @return expiresAt
+     */
+
+    @Nullable
+    public java.time.ZonedDateTime getExpiresAt() {
+        return this.expiresAt;
+    }
+
+    /**
      *  <p>State for the RecurringOrder in a custom workflow.</p>
      * @return state
      */
@@ -246,7 +270,7 @@ public class RecurringOrderDraftBuilder implements Builder<RecurringOrderDraft> 
         Objects.requireNonNull(cart, RecurringOrderDraft.class + ": cart is missing");
         Objects.requireNonNull(cartVersion, RecurringOrderDraft.class + ": cartVersion is missing");
         Objects.requireNonNull(startsAt, RecurringOrderDraft.class + ": startsAt is missing");
-        return new RecurringOrderDraftImpl(key, cart, cartVersion, startsAt, state, custom);
+        return new RecurringOrderDraftImpl(key, cart, cartVersion, startsAt, expiresAt, state, custom);
     }
 
     /**
@@ -254,7 +278,7 @@ public class RecurringOrderDraftBuilder implements Builder<RecurringOrderDraft> 
      * @return RecurringOrderDraft
      */
     public RecurringOrderDraft buildUnchecked() {
-        return new RecurringOrderDraftImpl(key, cart, cartVersion, startsAt, state, custom);
+        return new RecurringOrderDraftImpl(key, cart, cartVersion, startsAt, expiresAt, state, custom);
     }
 
     /**
@@ -276,6 +300,7 @@ public class RecurringOrderDraftBuilder implements Builder<RecurringOrderDraft> 
         builder.cart = template.getCart();
         builder.cartVersion = template.getCartVersion();
         builder.startsAt = template.getStartsAt();
+        builder.expiresAt = template.getExpiresAt();
         builder.state = template.getState();
         builder.custom = template.getCustom();
         return builder;