+++ changes.md
@@ -4,10 +4,13 @@
 <summary>Added Type(s)</summary>
 
 - added type `RecurringOrderDeletedMessage`
+- added type `RecurringOrderExpiresAtSetMessage`
 - added type `RecurringOrderDeletedMessagePayload`
+- added type `RecurringOrderExpiresAtSetMessagePayload`
 - added type `ProductSearchFacetResultStats`
 - added type `ProductSearchFacetStatsExpression`
 - added type `ProductSearchFacetStatsValue`
+- added type `RecurringOrderSetExpiresAtAction`
 </details>
 
 
@@ -23,6 +26,13 @@
 
 - added method `apiRoot.withProjectKey().recurringOrders().withId().delete()`
 - added method `apiRoot.withProjectKey().recurringOrders().withKey().delete()`
+</details>
+
+
+<details>
+<summary>Added Property(s)</summary>
+
+- added property `expiresAt` to type `RecurringOrderDraft`
 </details>
 
 **History changes**